pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "clicotea"
description = "Cross-Lingual Contextualised Token Embedding Alignment"
readme = "README.md"
dynamic = ["version"]
requires-python = ">=3.8"
authors = [
    { name = "Rémi Lebret", email = "remi@lebret.ch" },
    { name = "Yasmine Karoui", email = "karouiyasmine8@gmail.com" }
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python"
]
dependencies = [
    "numpy",
    "jsonlines",
    "torch",
    "googletrans==3.1.0a0",
    "transformers",
    "pytorch-ignite",
    "wandb",
    "tensorboardX",
    "salesforce-lavis",
]
license = "Apache-2.0 OR MIT"
keywords = [
    "machine learning",
    "deep learning",
    "pytorch",
    "transformers",
    "natural language processing"
]

[project.optional-dependencies]
test = [
    "coverage[toml]",
    "pytest",
    "pytest-cov",
    "pytest-mock",
]
notebook = [
    "jupyter",
    "ipython",
]
lint = ["pre-commit"]

[project.urls]
Documentation = "https://github.com/rlebret/clicotea#readme"
Issues = "https://github.com/rlebret/clicotea/issues"
Source = "https://github.com/rlebret/clicotea"

[tool.hatch.version]
path = "src/clicotea/__about__.py"

[tool.hatch.envs.default]
python = "3.10"
features = ["notebook", "lint"]
post-install-commands = [
    "pre-commit install",
]

[tool.hatch.envs.default.scripts]
translate = "python scripts/datasets/translate.py {args}"
alignment = "python scripts/alignment/run_token_alignment.py {args}"
clicotea = "python scripts/embeddings/train_clicotea.py {args}"
zeroshot = "python scripts/zero-shot/evaluate.py {args}"

[tool.hatch.envs.lint]
detached = true
features = ["lint"]

[tool.hatch.envs.lint.scripts]
run = "pre-commit run --all-files --show-diff-on-failure"

[tool.hatch.envs.test]
features = ["test"]

[tool.hatch.envs.test.scripts]
run-coverage = "pytest --cov-report=term-missing --cov-config=pyproject.toml --cov=clicotea --cov=tests {args}"
run = "run-coverage --no-cov {args}"

[tool.hatch.build.targets.sdist]
exclude = [
  "/.venv",
  "/.github",
  "/docs",
  "/results",
  "/scripts",
  "/data",
  "/docs",
]

[tool.hatch.build.targets.wheel]
packages = ["src/clicotea"]

[tool.ruff.per-file-ignores]
# Ignore lines that are too long in xbert.py
"src/clicotea/models/xbert.py"= ["E501"]
"tests/conftest.py"= ["E501"]
